power scheme, CPU core usage via PDH, class WinRing0_1_3RunTimeDynLinked
-CPU core usage via PDH:
 also for Windows < ver. 6
-power scheme:
 -can set and get the active power scheme for Windows < ver. 6
-class WinRing0_1_3RunTimeDynLinked:
 -adds implementation for ReadTSC() of base class I_CPUaccess
 -adds exception if WinRing0's initialization function "InitOpenLibSys()" fails
-Windows/DynFreqScalingThread: deleted IWbemServices sourcecode
 (is already contained in another file)

+++ Windows/PowerProf/PowerProfDynLinked.cpp
@@ -177,7 +177,8 @@ void PowerProfDynLinked::OutputAllPowerSchemes()
   }
 }
 
-BYTE PowerProfDynLinked::SetActivePowerScheme(
+//BYTE
+DWORD PowerProfDynLinked::SetActivePowerScheme(
   const std::wstring & r_stdwstrPowerSchemeName )
 {
   return mp_i_powerprofdynlinked->SetActivePowerScheme(
